Restructure the rest of tests
