+++ modules/cuba/src/test/resources/test_support/cuba_messages/foo/messages.properties
