+++ modules/cuba/src/test/resources/test_support/cuba_messages/messages_ru.properties
