--- modules/remoting/src/test/groovy/io/jmix/remoting/ClientTest.groovy
@@ -1,63 +1,0 @@
-/*
- * Copyright 2019 Haulmont.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.jmix.remoting
-
-import io.jmix.core.JmixCoreConfiguration
-import io.jmix.core.impl.ConfigStorage
-import io.jmix.data.JmixDataConfiguration
-import io.jmix.remoting.gateway.ClientConfigStorage
-import io.jmix.remoting.gateway.ServerConfigStorage
-import io.jmix.remoting.test.JmixRemotingTestConfiguration
-import io.jmix.remoting.test.TestService
-import org.springframework.context.ApplicationContext
-import org.springframework.test.context.ActiveProfiles
-import org.springframework.test.context.ContextConfiguration
-import spock.lang.Specification
-
-import javax.inject.Inject
-import java.lang.reflect.Proxy
-
-@ContextConfiguration(classes = [JmixCoreConfiguration, JmixDataConfiguration, JmixRemotingConfiguration,
-        JmixRemotingTestConfiguration])
-@ActiveProfiles(["remoting", "client"])
-class ClientTest extends Specification {
-
-    @Inject
-    ApplicationContext applicationContext
-
-    def "context has correct beans"() {
-
-        def configStorage = applicationContext.getBean(ConfigStorage.NAME)
-        def configStorageService = applicationContext.getBean(ServerConfigStorage.NAME)
-        def testService = applicationContext.getBean(TestService.class)
-
-        expect:
-
-        // client implementation
-        configStorage instanceof ClientConfigStorage
-
-        // service proxy
-        testService instanceof Proxy
-
-        // transport proxy
-        configStorageService instanceof Proxy
-
-        // no exports
-        !applicationContext.containsBean('/remoting/' + ServerConfigStorage.NAME)
-        !applicationContext.containsBean('/remoting/' + TestService.NAME)
-    }
-}
