+++ modules/remoting/src/test/java/test_support/TestServiceImpl.java
@@ -1,0 +1,29 @@
+/*
+ * Copyright 2019 Haulmont.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package test_support;
+
+import io.jmix.remoting.annotation.Remote;
+import org.springframework.stereotype.Component;
+
+@Component(TestService.NAME)
+@Remote
+public class TestServiceImpl implements TestService {
+    @Override
+    public String echo(String input) {
+        return input;
+    }
+}
